requirements.txt
beautifulsoup4==4.12.3
chardet==5.2.0
sanic==23.12.1
sanic-ext==23.12.0
uvicorn==0.27.0
colorama==0.4.6
nslookup~=1.7.0
python-jose==3.3.0
pyDes==2.0.1
tldextract==5.1.1
python-whois==0.8.0
python-docx==1.1.0
aiomysql==0.2.0
aiohttp==3.9.5
urllib3~=1.26.18
bs4~=0.0.2
dnspython~=2.6.1
DBUtils~=3.1.0